[Oztechan/CCC#2259] Move premium related models and utils into :client:viewmodel:premium module
