+++ client/viewmodel/premium/client-viewmodel-premium.gradle.kts
@@ -21,6 +21,7 @@ kotlin {
                     implementation(koinCore)
                     implementation(kermit)
                     implementation(coroutines)
+                    implementation(kotlinXDateTime)
                 }
 
                 Modules.Client.Core.apply {
