+++ client/viewmodel/premium/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/premium/PremiumSEED.kt
@@ -1,9 +1,9 @@
 package com.oztechan.ccc.client.viewmodel.premium
 
-import com.oztechan.ccc.client.core.shared.model.PremiumType
 import com.oztechan.ccc.client.core.viewmodel.BaseEffect
 import com.oztechan.ccc.client.core.viewmodel.BaseEvent
 import com.oztechan.ccc.client.core.viewmodel.BaseState
+import com.oztechan.ccc.client.viewmodel.premium.model.PremiumType
 
 // State
 data class PremiumState(
