+++ client/viewmodel/premium/src/commonMain/kotlin/com/oztechan/ccc/client/viewmodel/premium/PremiumViewModel.kt
@@ -6,10 +6,6 @@ package com.oztechan.ccc.client.viewmodel.premium
 
 import co.touchlab.kermit.Logger
 import com.github.submob.scopemob.whether
-import com.oztechan.ccc.client.core.shared.model.OldPurchase
-import com.oztechan.ccc.client.core.shared.model.PremiumData
-import com.oztechan.ccc.client.core.shared.model.PremiumType
-import com.oztechan.ccc.client.core.shared.util.calculatePremiumEnd
 import com.oztechan.ccc.client.core.shared.util.isNotPassed
 import com.oztechan.ccc.client.core.shared.util.nowAsLong
 import com.oztechan.ccc.client.core.viewmodel.BaseData
@@ -17,6 +13,10 @@ import com.oztechan.ccc.client.core.viewmodel.BaseSEEDViewModel
 import com.oztechan.ccc.client.core.viewmodel.util.launchIgnored
 import com.oztechan.ccc.client.core.viewmodel.util.update
 import com.oztechan.ccc.client.storage.app.AppStorage
+import com.oztechan.ccc.client.viewmodel.premium.model.OldPurchase
+import com.oztechan.ccc.client.viewmodel.premium.model.PremiumData
+import com.oztechan.ccc.client.viewmodel.premium.model.PremiumType
+import com.oztechan.ccc.client.viewmodel.premium.util.calculatePremiumEnd
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asSharedFlow
